chore: Move `scanning` module into `startup`
