+++ mobile/src/modules/startup/scanning/ScanningProgress.ts
