+++ mobile/src/modules/startup/scanning/core/artwork.ts
