+++ mobile/src/modules/startup/scanning/core/audio.ts
