+++ mobile/src/modules/startup/scanning/core/cleanup.ts
