+++ mobile/src/modules/startup/scanning/core/rescan.ts
