+++ mobile/src/modules/startup/scanning/useScanning.ts
@@ -7,9 +7,9 @@ import { useCallback, useEffect, useState } from "react";
 import { Resynchronize } from "~/stores/Playback/actions";
 import { preferenceStore } from "~/stores/Preference/store";
 
-import { findAndSaveArtwork } from "../helpers/artwork";
-import { findAndSaveAudio } from "../helpers/audio";
-import { AppCleanUp } from "../helpers/cleanup";
+import { findAndSaveArtwork } from "./core/artwork";
+import { findAndSaveAudio } from "./core/audio";
+import { AppCleanUp } from "./core/cleanup";
 
 import { Stopwatch } from "~/utils/debug";
