+++ mobile/src/modules/startup/screens/OnboardingConfigurationView.tsx
